requirements/local.txt
-r base.txt
coverage==3.7.1
Sphinx==1.3b3
pudb==2015.2
django-extensions==1.5.1
django-debug-toolbar==1.3.0
